pyproject.toml
[project]
name = "videomatic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.6",
    "google-cloud-aiplatform>=1.136.0",
    "google-cloud-texttospeech>=2.34.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "moviepy>=2.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
